♻️ REFACTOR: Asset compilation and dev workflow
This refactor follows development of the workflow in https://github.com/executablebooks/sphinx-panels and https://github.com/executablebooks/web-compile.
It fixes a number of pressing issues (particularly see https://github.com/executablebooks/jupyter-book/issues/921):

- The package is using a SCSS compiler (pyScss) which is not well maintained and has some important un-adressed issues (e.g. it can't handle compiling CSS custom variables)
- Currently we are writing CSS files into the installed folder (not good at all!)
- The CSS/JS is not hashed, and so when people update their sphinx-book-theme, they may not always get the up-to-date files (due to caching)

The approach here is essentially very similar to pydata-sphinx-theme. But instead of using webpack to compile the SCSS/JS we are using the web-compile pre-commit hook. This limits the development complexity of having to work with an additional nodejs environment, and allows everything to run *via* tox.

As previously discussed, both of these approaches (including the compiled/hashed files in the repo) do suffer one drawback: that merge conflicts can arise if not making PRs to the latest commit on the `master` branch.
However, this use-case has literally never come up yet, so I'd rather accept this con for now, in return for the more important fixes/pros.
Note, there is talk of eventually doing this compilation at build time (see https://github.com/executablebooks/meta/issues/137), but I'm not satisfied yet that there is an ideal solution (and in any case this is a step in the direction).

As discussed previously I moved from nox to tox, since tox is (a) the more mature/used package and (b) used across other EBP packages.