Fix TOC hiding for "tall" full-width elements
I have a full-width image that is also long in the vertical dimension. As I scrolled down, the image would come into view from below, and when the top margin overlapped the "TOC zone", the TOC would be hidden. That's pretty nifty, thanks! But as the top margin disappeared above, the TOC was displayed over the middle of the image. Then as the bottom of the image overlapped the "TOC zone", the TOC would be hidden again. This commit fixes the `removeToc` boolean to include the "tall" full-width element case.