Escape colon in JupyterHub link to repo
I've run into a problem using The Littlest JupyterHub with current
`sphinx-book-theme`. At the moment the generated URLs for JupyterHub are
of form:

```
<hub-url>/hub/user-redirect/git-pull?repo=https://<repo>&urlpath=tree/<nb_fname>
```

These work fine on a modern Kubernetes JupyterHub, I guess because of
a recent version of JupyterHub responding to the query.

But, on current The Littlest JupyterHub, still on JupyterHub 1.5.0, this
causes Nbgitpuller to stall if the user is not already logged in, with
log messages of form:

```
Disallowing redirect outside JupyterHub:
'/hub/user-redirect/git-pull?repo=https://github.com/<repo>&urlpath=tree/<nb_fname>'.
```

Once logged in, retrying the same link works correctly.

This can be fixed by escaping the `:` in the Nbgitpuller URL with a
`%3A`, in this PR.