MAINT: remove incorrect comment from sphinx pin.
The comment here was out-of-date (sphinx 5 has been released)
which made it unclear whether the pin was intentional or an
error. From the discussion in gh-583, it's now clear the pin
is intentional, so removing the comment for clarity.

Alternatively, the comment could be updated in a way that won't
be obsoleted by future sphinx releases - wording suggestions
welcome.